+++ src/main/java/net/frozenblock/lib/FrozenLibClient.java
@@ -24,6 +24,7 @@ import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
 import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientWorldEvents;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayConnectionEvents;
 import net.fabricmc.fabric.api.resource.ResourceManagerHelper;
+import net.frozenblock.lib.block.client.render_type.impl.BlockRenderTypeOverwriteManager;
 import net.frozenblock.lib.block.sound.impl.BlockSoundTypeManager;
 import net.frozenblock.lib.cape.client.impl.ClientCapeData;
 import net.frozenblock.lib.config.frozenlib_config.FrozenLibConfig;
@@ -73,6 +74,7 @@ public final class FrozenLibClient implements ClientModInitializer {
 
 		var resourceLoader = ResourceManagerHelper.get(PackType.CLIENT_RESOURCES);
 		resourceLoader.registerReloadListener(BlockSoundTypeManager.INSTANCE);
+		resourceLoader.registerReloadListener(BlockRenderTypeOverwriteManager.INSTANCE);
 
 		FrozenClientEntrypoint.EVENT.invoker().init(); // also includes dev init
 	}