+++ src/main/java/net/frozenblock/lib/block/client/render_type/impl/BlockRenderTypeOverwrite.java
@@ -1,0 +1,86 @@
+/*
+ * Copyright (C) 2025 FrozenBlock
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package net.frozenblock.lib.block.client.render_type.impl;
+
+import com.mojang.serialization.Codec;
+import net.fabricmc.api.EnvType;
+import net.fabricmc.api.Environment;
+import net.minecraft.client.renderer.RenderType;
+import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.util.StringRepresentable;
+import net.minecraft.world.level.block.Block;
+import org.jetbrains.annotations.ApiStatus;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.Optional;
+import java.util.function.Supplier;
+
+@ApiStatus.Internal
+@Environment(EnvType.CLIENT)
+public class BlockRenderTypeOverwrite {
+	private final ResourceLocation blockName;
+	private final RenderTypeOverwrite renderTypeOverwrite;
+
+	public BlockRenderTypeOverwrite(ResourceLocation blockName, RenderTypeOverwrite renderTypeOverwrite) {
+		this.blockName = blockName;
+		this.renderTypeOverwrite = renderTypeOverwrite;
+	}
+
+	public ResourceLocation getBlockName() {
+		return this.blockName;
+	}
+
+	public RenderTypeOverwrite getRenderTypeOverwrite() {
+		return this.renderTypeOverwrite;
+	}
+
+	public RenderType getRenderType() {
+		return this.getRenderTypeOverwrite().get();
+	}
+
+	public Optional<Block> getBlock() {
+		return BuiltInRegistries.BLOCK.getOptional(this.blockName);
+	}
+
+	public enum RenderTypeOverwrite implements StringRepresentable {
+		SOLID("solid", RenderType::solid),
+		CUTOUT_MIPPED("cutout_mipped", RenderType::cutoutMipped),
+		CUTOUT("cutout", RenderType::cutout),
+		TRANSLUCENT("translucent", RenderType::translucent),
+		TRIPWIRE("tripwire", RenderType::tripwire);
+		public static final Codec<RenderTypeOverwrite> CODEC = StringRepresentable.fromEnum(RenderTypeOverwrite::values);
+
+		private final String name;
+		private final Supplier<RenderType> renderTypeSupplier;
+
+		RenderTypeOverwrite(String name, Supplier<RenderType> renderTypeSupplier) {
+			this.name = name;
+			this.renderTypeSupplier = renderTypeSupplier;
+		}
+
+		public RenderType get() {
+			return this.renderTypeSupplier.get();
+		}
+
+		@Override
+		public @NotNull String getSerializedName() {
+			return this.name;
+		}
+	}
+}